www/react: Move useTopbarItems hook to buildbot-ui package
