+++ www/react-ui/src/main.ts
@@ -31,9 +31,11 @@ export * from "./contexts/Config";
 export * from "./contexts/Time";
 
 export * from "./stores/TimeStore";
+export * from "./stores/TopbarStore";
 
 export * from "./util/FavIcon";
 export * from "./util/Moment";
 export * from "./util/React";
 export * from "./util/StepUrls";
 export * from "./util/TagFilterManager";
+export * from "./util/Topbar";
