+++ www/react-ui/src/stores/TopbarStore.ts
@@ -16,8 +16,11 @@
 */
 
 import {action, IObservableArray, makeObservable, observable} from "mobx";
-import {TopbarItem} from "../components/Topbar/Topbar";
-import {useEffect} from "react";
+
+export type TopbarItem = {
+  route: string | null,
+  caption: string
+}
 
 export class TopbarStore {
   items: IObservableArray<TopbarItem> = observable<TopbarItem>([]);
@@ -33,16 +36,4 @@ export class TopbarStore {
   @action clearItems() {
     this.items.clear();
   }
-}
-
-export function useTopbarItems(store: TopbarStore, items: TopbarItem[]) {
-  useEffect(() => {
-    store.setItems(items);
-  }, [store, items]);
-
-  // We only want to clear the items once, thus the useEffect hook is split into two parts, one
-  // for updates, one for eventual cleanup when navigating out of view.
-  useEffect(() => {
-    return () => store.setItems([]);
-  }, [store])
 }
